% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{sd_img2img}
\alias{sd_img2img}
\title{Generate images with img2img}
\usage{
sd_img2img(
  ctx,
  prompt,
  init_image,
  negative_prompt = "",
  width = NULL,
  height = NULL,
  sample_method = SAMPLE_METHOD$EULER,
  sample_steps = 20L,
  cfg_scale = 7,
  seed = 42L,
  batch_count = 1L,
  scheduler = SCHEDULER$DISCRETE,
  clip_skip = -1L,
  strength = 0.75,
  eta = 0
)
}
\arguments{
\item{ctx}{SD context created by \code{\link{sd_ctx}}}

\item{prompt}{Text prompt describing desired image}

\item{init_image}{Init image in sd_image format. Use \code{\link{sd_load_image}}
to load from file.}

\item{negative_prompt}{Negative prompt (default "")}

\item{width}{Image width in pixels (default 512)}

\item{height}{Image height in pixels (default 512)}

\item{sample_method}{Sampling method (see \code{SAMPLE_METHOD})}

\item{sample_steps}{Number of sampling steps (default 20)}

\item{cfg_scale}{Classifier-free guidance scale (default 7.0)}

\item{seed}{Random seed (-1 for random)}

\item{batch_count}{Number of images to generate (default 1)}

\item{scheduler}{Scheduler type (see \code{SCHEDULER})}

\item{clip_skip}{Number of CLIP layers to skip (-1 = auto)}

\item{strength}{Denoising strength (0.0 = no change, 1.0 = full denoise, default 0.75)}

\item{eta}{Eta parameter for DDIM-like samplers}
}
\value{
List of SD images
}
\description{
Generate images with img2img
}
